requirements.txt
numpy
scipy
scikit-learn
umap-learn
matplotlib
seaborn
plotly
streamlit
torch
torchvision
transformers
diffusers
accelerate
torchvision
streamlit
plotly
opencv-python
Pillow
jsonlines
tqdm
pyyaml
types-PyYAML
einops
